wgan/requirements.txt
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
appdirs==1.4.3
asn1crypto @ file:///tmp/build/80754af9/asn1crypto_1596577642040/work
astroid==2.3.3
astropy==4.0.2
async-generator==1.10
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4 @ file:///tmp/build/80754af9/beautifulsoup4_1601924105527/work
bitarray @ file:///C:/ci/bitarray_1603203287144/work
bkcharts==0.2
bleach==3.1.4
bokeh @ file:///C:/ci/bokeh_1603297932359/work
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1600699250966/work
chardet==3.0.4
click==7.1.2
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
clyent==1.2.2
colorama==0.4.3
comtypes==1.1.7
conda==4.9.1
conda-build==3.18.9
conda-package-handling @ file:///C:/ci/conda-package-handling_1603018162806/work
conda-verify==3.4.2
contextlib2==0.6.0.post1
cryptography @ file:///C:/ci/cryptography_1601046913206/work
cycler==0.10.0
Cython @ file:///C:/ci/cython_1594834055134/work
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1602083700509/work
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distributed @ file:///C:/ci/distributed_1602065865185/work
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.2
fsspec @ file:///tmp/build/80754af9/fsspec_1602684995936/work
future==0.18.2
gevent @ file:///C:/ci/gevent_1601397624717/work
glob2==0.7
greenlet @ file:///C:/ci/greenlet_1600874076725/work
h5py==2.10.0
HeapDict==1.0.1
html5lib @ file:///tmp/build/80754af9/html5lib_1593446221756/work
hypothesis @ file:///tmp/build/80754af9/hypothesis_1604100955198/work
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
imagesize==1.2.0
importlib-metadata==1.6.0
iniconfig @ file:///tmp/build/80754af9/iniconfig_1602780191262/work
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-http-over-ws==0.0.8
jupyterlab==2.2.6
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
keyring @ file:///C:/ci/keyring_1601488971022/work
kiwisolver @ file:///C:/ci/kiwisolver_1603996595157/work
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.34.0
locket==0.2.0
lxml @ file:///C:/ci/lxml_1603216364379/work
MarkupSafe==1.1.1
matplotlib @ file:///C:/ci/matplotlib-base_1603356257853/work
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
more-itertools @ file:///tmp/build/80754af9/more-itertools_1598884071673/work
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
navigator-updater==0.2.1
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert==5.6.1
nbformat==5.0.4
nbopen==0.6
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1601499549014/work
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
nltk @ file:///tmp/build/80754af9/nltk_1592496090529/work
nose @ file:///C:/ci/nose_1594377973236/work
notebook==6.0.3
numba==0.51.2
numexpr==2.7.1
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
numpydoc @ file:///tmp/build/80754af9/numpydoc_1594166760263/work
olefile==0.46
opencv-python==4.4.0.46
openpyxl @ file:///tmp/build/80754af9/openpyxl_1598113097404/work
packaging==20.4
pandas @ file:///C:/ci/pandas_1602088205210/work
pandocfilters==1.4.2
parso==0.6.2
partd==1.1.0
path @ file:///C:/ci/path_1596940809557/work
pathlib2 @ file:///C:/ci/pathlib2_1594381094851/work
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1603821929285/work
pkginfo==1.6.0
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psutil @ file:///C:/ci/psutil_1598352273540/work
py @ file:///tmp/build/80754af9/py_1593446248552/work
pycodestyle==2.6.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycrypto==2.6.1
pycurl==7.43.0.6
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.4.4
pyodbc===4.0.0-unsupported
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.16.0
PySocks @ file:///C:/ci/pysocks_1594394709107/work
pytest==0.0.0
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus @ file:///tmp/build/80754af9/pytest-doctestplus_1598624048543/work
pytest-openfiles==0.5.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
pytz==2020.1
PyWavelets @ file:///C:/ci/pywavelets_1601658407053/work
pywin32==227
pywin32-ctypes @ file:///C:/ci/pywin32-ctypes_1594392691209/work
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
QtAwesome @ file:///tmp/build/80754af9/qtawesome_1602272867890/work
qtconsole==4.7.2
QtPy==1.9.0
regex @ file:///C:/ci/regex_1602782614108/work
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
rope @ file:///tmp/build/80754af9/rope_1602264064449/work
ruamel-yaml==0.15.87
scikit-image==0.17.2
scikit-learn @ file:///C:/ci/scikit-learn_1598376983131/work
scipy @ file:///C:/ci/scipy_1597686737426/work
seaborn @ file:///tmp/build/80754af9/seaborn_1600553570093/work
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch @ file:///tmp/build/80754af9/singledispatch_1602523705405/work
six==1.14.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx @ file:///tmp/build/80754af9/sphinx_1597428793432/work
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport @ file:///tmp/build/80754af9/sphinxcontrib-websupport_1597081412696/work
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy @ file:///C:/ci/sqlalchemy_1603818968375/work
statsmodels==0.12.0
sympy @ file:///C:/ci/sympy_1597083371142/work
tables==3.6.1
tblib @ file:///tmp/build/80754af9/tblib_1597928476713/work
terminado==0.8.3
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tifffile==2020.10.1
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
toolz @ file:///tmp/build/80754af9/toolz_1601054250827/work
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1602185206534/work
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
unicodecsv==0.14.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
virtualenv==20.0.17
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter @ file:///tmp/build/80754af9/xlsxwriter_1602692860603/work
xlwings==0.20.8
xlwt==1.3.0
xmltodict==0.12.0
zict==2.0.0
zipp==3.1.0
zope.event==4.5.0
zope.interface @ file:///C:/ci/zope.interface_1602002496598/work
